HS-926: clean up classes
