--- metrics-processor/flows/src/main/java/org/opennms/horizon/flows/classification/internal/validation/GroupValidator.java
@@ -1,58 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2018-2018 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.horizon.flows.classification.internal.validation;
-
-import org.opennms.horizon.flows.classification.error.ErrorContext;
-import org.opennms.horizon.flows.classification.error.Errors;
-import org.opennms.horizon.flows.classification.exception.ClassificationException;
-import org.opennms.horizon.flows.classification.persistence.api.ClassificationRuleDao;
-import org.opennms.horizon.flows.classification.persistence.api.Group;
-import org.opennms.horizon.flows.classification.persistence.api.Rule;
-
-import java.util.Objects;
-
-public class GroupValidator {
-
-    private final ClassificationRuleDao ruleDao;
-
-    public GroupValidator(final ClassificationRuleDao ruleDao) {
-        this.ruleDao = Objects.requireNonNull(ruleDao);
-    }
-
-    // verify if adding the given rule to the given group would not result in any conflict(s)
-    public void validate(Group group, Rule potentialNewRule) {
-        if (potentialNewRule != null) {
-            Objects.requireNonNull(group);
-            final Rule existingRule = ruleDao.findByDefinition(potentialNewRule, group);
-            if (existingRule != null) {
-                throw new ClassificationException(ErrorContext.Name, Errors.GROUP_DUPLICATE_RULE);
-            }
-        }
-    }
-}
