--- metrics-processor/flows/src/main/java/org/opennms/horizon/flows/integration/Directional.java
@@ -1,74 +1,0 @@
-/*******************************************************************************
- * This file is part of OpenNMS(R).
- *
- * Copyright (C) 2017 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2017 The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
- *
- * OpenNMS(R) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published
- * by the Free Software Foundation, either version 3 of the License,
- * or (at your option) any later version.
- *
- * OpenNMS(R) is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with OpenNMS(R).  If not, see:
- *      http://www.gnu.org/licenses/
- *
- * For more information contact:
- *     OpenNMS(R) Licensing <license@opennms.org>
- *     http://www.opennms.org/
- *     http://www.opennms.com/
- *******************************************************************************/
-
-package org.opennms.horizon.flows.integration;
-
-import java.util.Objects;
-
-/**
- * Used to associate some object with a direction.
- */
-public class Directional<T> {
-    private final T value;
-    private final boolean isIngress;
-
-    public Directional(T value, boolean isIngress) {
-        this.value = Objects.requireNonNull(value);
-        this.isIngress = isIngress;
-    }
-
-    public T getValue() {
-        return value;
-    }
-
-    public boolean isIngress() {
-        return isIngress;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Directional<?> that = (Directional<?>) o;
-        return isIngress == that.isIngress &&
-                Objects.equals(value, that.value);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(value, isIngress);
-    }
-
-    @Override
-    public String toString() {
-        return "Directional{" +
-                "value=" + value +
-                ",ingress=" + isIngress +
-                '}';
-    }
-}
