+++ metrics-processor/flows/src/test/java/org/opennms/horizon/flows/classification/csv/CsvServiceTest.java
@@ -31,9 +31,6 @@ package org.opennms.horizon.flows.classification.csv;
 import com.google.common.collect.Lists;
 import org.junit.Before;
 import org.junit.Test;
-import org.opennms.horizon.flows.classification.FilterService;
-import org.opennms.horizon.flows.classification.exception.InvalidFilterException;
-import org.opennms.horizon.flows.classification.internal.CachingFilterService;
 import org.opennms.horizon.flows.classification.internal.csv.CsvServiceImpl;
 import org.opennms.horizon.flows.classification.internal.validation.RuleValidator;
 import org.opennms.horizon.flows.classification.persistence.api.Group;
